apps/vision/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pillow>=11.0.0
numpy>=2.0.0
pydantic>=2.10.0
python-multipart>=0.0.18
